+++ lib/tty.js
@@ -27,14 +27,7 @@ const { TTY, isTTY } = process.binding('tty_wrap');
 const errors = require('internal/errors');
 const { ERR_INVALID_FD } = errors.codes;
 const readline = require('readline');
-const { release } = require('os');
-
-const OSRelease = release().split('.');
-
-const COLORS_2 = 1;
-const COLORS_16 = 4;
-const COLORS_256 = 8;
-const COLORS_16m = 24;
+const { getColorDepth } = require('internal/tty');
 
 function isatty(fd) {
   return Number.isInteger(fd) && fd >= 0 && isTTY(fd);
@@ -108,70 +101,7 @@ inherits(WriteStream, net.Socket);
 
 WriteStream.prototype.isTTY = true;
 
-WriteStream.prototype.getColorDepth = function(env = process.env) {
-  if (env.NODE_DISABLE_COLORS || env.TERM === 'dumb' && !env.COLORTERM) {
-    return COLORS_2;
-  }
-
-  if (process.platform === 'win32') {
-    // Windows 10 build 10586 is the first Windows release that supports 256
-    // colors. Windows 10 build 14931 is the first release that supports
-    // 16m/TrueColor.
-    if (+OSRelease[0] >= 10) {
-      const build = +OSRelease[2];
-      if (build >= 14931)
-        return COLORS_16m;
-      if (build >= 10586)
-        return COLORS_256;
-    }
-
-    return COLORS_16;
-  }
-
-  if (env.TMUX) {
-    return COLORS_256;
-  }
-
-  if (env.CI) {
-    if ('TRAVIS' in env || 'CIRCLECI' in env || 'APPVEYOR' in env ||
-      'GITLAB_CI' in env || env.CI_NAME === 'codeship') {
-      return COLORS_256;
-    }
-    return COLORS_2;
-  }
-
-  if ('TEAMCITY_VERSION' in env) {
-    return /^(9\.(0*[1-9]\d*)\.|\d{2,}\.)/.test(env.TEAMCITY_VERSION) ?
-      COLORS_16 : COLORS_2;
-  }
-
-  switch (env.TERM_PROGRAM) {
-    case 'iTerm.app':
-      if (!env.TERM_PROGRAM_VERSION ||
-        /^[0-2]\./.test(env.TERM_PROGRAM_VERSION)) {
-        return COLORS_256;
-      }
-      return COLORS_16m;
-    case 'HyperTerm':
-    case 'Hyper':
-    case 'MacTerm':
-      return COLORS_16m;
-    case 'Apple_Terminal':
-      return COLORS_256;
-  }
-
-  if (env.TERM) {
-    if (/^xterm-256/.test(env.TERM))
-      return COLORS_256;
-    if (/^screen|^xterm|^vt100|color|ansi|cygwin|linux/i.test(env.TERM))
-      return COLORS_16;
-  }
-
-  if (env.COLORTERM)
-    return COLORS_16;
-
-  return COLORS_2;
-};
+WriteStream.prototype.getColorDepth = getColorDepth;
 
 WriteStream.prototype._refreshSize = function() {
   const oldCols = this.columns;
